feat: add ProjectsView and CertificatesView components with data fetching and management functionality
