+++ src/App.jsx
@@ -7,6 +7,8 @@ import AnimatedSuspenseWrapper from "./components/utils/AnimatedSuspenseWrapper"
 import CircleLoader from "./components/utils/CircleLoader";
 import Login from "./components/Login";
 import { AuthProvider } from "./context/AuthContext";
+import ProjectsView from "./components/utils/ProjectsView";
+import CertificatesView from "./components/utils/CertificatesView";
 
 const DashBoard = lazy(() => import("./components/DashBoard"));
 
@@ -22,6 +24,8 @@ function App() {
               <Route path="/admin" element={<PrivateLayout />}>
                 <Route index element={<Login />} />
                 <Route path="dashboard" element={<DashBoard />}/>
+                <Route path="/admin/projects" element={<ProjectsView />} />
+                <Route path="/admin/certificates" element={<CertificatesView />} />
               </Route>
             </Routes>
           </AnimatedSuspenseWrapper>
